chore: store: db: reuse connection pool and change configuration
The pool object is meant to be reused and since we're initializing the
store as part of a long running process, we should save the pool object
and pass it around/use it.
Have a proper db url directly in configuration files instead of using
environment variables.
