+++ db/src/lib.rs
@@ -20,10 +20,7 @@ where
     T: diesel::r2d2::R2D2Connection + 'static,
 {
     /// Gets a connection pool
-    fn get_conn_pool() -> Pool<ConnectionManager<T>>;
-
-    /// Gets a connection to the db
-    fn get_connection() -> T;
+    fn get_conn_pool(url: String) -> Pool<ConnectionManager<T>>;
 
     /// Inserts an OV
     fn insert_ov(ov: &OV, ttl: Option<i64>, conn: &mut T) -> Result<()>;
@@ -51,10 +48,7 @@ where
     T: diesel::r2d2::R2D2Connection + 'static,
 {
     /// Gets a connection pool
-    fn get_conn_pool() -> Pool<ConnectionManager<T>>;
-
-    /// Gets a connection to the db
-    fn get_connection() -> T;
+    fn get_conn_pool(url: String) -> Pool<ConnectionManager<T>>;
 
     /// Inserts an OV
     fn insert_ov(ov: &OV, to2: Option<bool>, to0: Option<i64>, conn: &mut T) -> Result<()>;
@@ -99,10 +93,7 @@ where
     T: diesel::r2d2::R2D2Connection + 'static,
 {
     /// Gets a connection pool
-    fn get_conn_pool() -> Pool<ConnectionManager<T>>;
-
-    /// Gets a connection to the db
-    fn get_connection() -> T;
+    fn get_conn_pool(url: String) -> Pool<ConnectionManager<T>>;
 
     /// Inserts an OV
     fn insert_ov(ov: &StoredItem, guid: &str, ttl: Option<i64>, conn: &mut T) -> Result<()>;
