+++ db/src/postgres.rs
@@ -10,8 +10,6 @@ use diesel::r2d2::ConnectionManager;
 use diesel::r2d2::Pool;
 use diesel::PgConnection;
 
-use std::env;
-
 use anyhow::Result;
 
 use super::models::{ManufacturerOV, NewOwnerOV, NewRendezvousOV, OwnerOV, RendezvousOV};
@@ -22,15 +20,7 @@ use fdo_data_formats::Serializable;
 pub struct PostgresManufacturerDB {}
 
 impl DBStoreManufacturer<PgConnection> for PostgresManufacturerDB {
-    fn get_connection() -> PgConnection {
-        let database_url = env::var("POSTGRES_MANUFACTURER_DATABASE_URL")
-            .expect("POSTGRES_MANUFACTURER_DATABASE_URL must be set");
-        PgConnection::establish(&database_url).expect("Error connecting to database")
-    }
-
-    fn get_conn_pool() -> Pool<ConnectionManager<PgConnection>> {
-        let database_url = env::var("POSTGRES_MANUFACTURER_DATABASE_URL")
-            .expect("POSTGRES_MANUFACTURER_DATABASE_URL must be set");
+    fn get_conn_pool(database_url: String) -> Pool<ConnectionManager<PgConnection>> {
         let manager = ConnectionManager::<PgConnection>::new(database_url);
         Pool::builder()
             .test_on_check_out(true)
@@ -90,15 +80,7 @@ impl DBStoreManufacturer<PgConnection> for PostgresManufacturerDB {
 pub struct PostgresOwnerDB {}
 
 impl DBStoreOwner<PgConnection> for PostgresOwnerDB {
-    fn get_connection() -> PgConnection {
-        let database_url = env::var("POSTGRES_OWNER_DATABASE_URL")
-            .expect("POSTGRES_OWNER_DATABASE_URL must be set");
-        PgConnection::establish(&database_url).expect("Error connecting to database")
-    }
-
-    fn get_conn_pool() -> Pool<ConnectionManager<PgConnection>> {
-        let database_url = env::var("POSTGRES_OWNER_DATABASE_URL")
-            .expect("POSTGRES_OWNER_DATABASE_URL must be set");
+    fn get_conn_pool(database_url: String) -> Pool<ConnectionManager<PgConnection>> {
         let manager = ConnectionManager::<PgConnection>::new(database_url);
         Pool::builder()
             .test_on_check_out(true)
@@ -208,15 +190,7 @@ impl DBStoreOwner<PgConnection> for PostgresOwnerDB {
 pub struct PostgresRendezvousDB {}
 
 impl DBStoreRendezvous<PgConnection> for PostgresRendezvousDB {
-    fn get_connection() -> PgConnection {
-        let database_url = env::var("POSTGRES_RENDEZVOUS_DATABASE_URL")
-            .expect("POSTGRES_RENDEZVOUS_DATABASE_URL must be set");
-        PgConnection::establish(&database_url).expect("Error connecting to database")
-    }
-
-    fn get_conn_pool() -> Pool<ConnectionManager<PgConnection>> {
-        let database_url = env::var("POSTGRES_RENDEZVOUS_DATABASE_URL")
-            .expect("POSTGRES_RENDEZVOUS_DATABASE_URL must be set");
+    fn get_conn_pool(database_url: String) -> Pool<ConnectionManager<PgConnection>> {
         let manager = ConnectionManager::<PgConnection>::new(database_url);
         Pool::builder()
             .test_on_check_out(true)
