+++ test/setup.sh
@@ -66,7 +66,7 @@ sudo systemctl start libvirtd
 sudo virsh list --all > /dev/null
 
 # Set a customized dnsmasq configuration for libvirt so we always get the
-# same address on bootup.
+# same address on boot.
 greenprint "💡 Setup libvirt network"
 sudo tee /tmp/integration.xml > /dev/null << EOF
 <network xmlns:dnsmasq='http://libvirt.org/schemas/network/dnsmasq/1.0'>
